+++ lua/render-markdown/ui.lua
@@ -1,5 +1,4 @@
 local logger = require('render-markdown.logger')
-local profiler = require('render-markdown.profiler')
 local state = require('render-markdown.state')
 local util = require('render-markdown.util')
 
@@ -31,13 +30,7 @@ end
 function M.schedule_render(buf, parse)
     local mode = vim.fn.mode(true)
     vim.schedule(function()
-        if state.config.profile then
-            profiler.profile(buf, function()
-                return M.render(buf, mode, parse)
-            end)
-        else
-            M.render(buf, mode, parse)
-        end
+        M.render(buf, mode, parse)
     end)
 end
 
@@ -45,12 +38,11 @@ end
 ---@param buf integer
 ---@param mode string
 ---@param parse boolean
----@return 'invalid'|'disable'|'parsed'|'movement'
 function M.render(buf, mode, parse)
     -- Check that buffer and associated window are valid
     local win = util.buf_to_win(buf)
     if not vim.api.nvim_buf_is_valid(buf) or not vim.api.nvim_win_is_valid(win) then
-        return 'invalid'
+        return
     end
     vim.api.nvim_buf_clear_namespace(buf, M.namespace, 0, -1)
 
@@ -59,7 +51,6 @@ function M.render(buf, mode, parse)
         for name, value in pairs(state.config.win_options) do
             util.set_win(win, name, value.default)
         end
-        return 'disable'
     else
         -- Set window options to rendered & perform render
         for name, value in pairs(state.config.win_options) do
@@ -84,12 +75,6 @@ function M.render(buf, mode, parse)
                 mark.opts.strict = parsed
                 vim.api.nvim_buf_set_extmark(buf, M.namespace, mark.start_row, mark.start_col, mark.opts)
             end
-        end
-
-        if parsed then
-            return 'parsed'
-        else
-            return 'movement'
         end
     end
 end