+++ tests/box_dash_quote_spec.lua
@@ -1,4 +1,5 @@
 ---@module 'luassert'
+
 local util = require('tests.util')
 
 describe('box_dash_quote.md', function()